+++ packages/control/algorithm/filter_chargepoints_test.py
@@ -10,7 +10,7 @@ from control.chargemode import Chargemode
 from control.chargepoint.chargepoint import Chargepoint, ChargepointData
 from control.chargepoint.chargepoint_data import Log, Set
 from control.chargepoint.control_parameter import ControlParameter
-from control.counter_all import CounterAll
+from control.counter_all.counter_all import CounterAll
 from control.ev.ev import Ev, EvData, Get
