+++ packages/control/counter_all/counter_home_consumption_test.py
@@ -4,7 +4,7 @@ import pytest
 
 from control import data
 from packages.conftest import hierarchy_hc_counter, hierarchy_standard, hierarchy_hybrid, hierarchy_nested
-from control.counter_all import CounterAll
+from control.counter_all.counter_all import CounterAll
 from modules.common.fault_state import FaultStateLevel
